Revert "Lv2. SearchBar 연동 && JSON 기반 통화코드-국가명 매핑"
