+++ ExchangesApp/Assets.xcassets/AccentColor.colorset/Contents.json
