+++ ExchangesApp/Assets.xcassets/AppIcon.appiconset/Contents.json
