+++ ExchangesApp/Assets.xcassets/Contents.json
