--- ExchangesApp/Sources/Model/ExchangeRateDisplayModel.swift
@@ -1,14 +1,0 @@
-//
-//  ExchangeRateDisplayModel.swift
-//  ExchangesApp
-//
-//  Created by 노가현 on 7/9/25.
-//
-
-import Foundation
-
-struct ExchangeRateDisplayModel {
-    let currencyCode: String // 통화 코드
-    let countryName: String // 통화의 국가 이름
-    let rate: Double // 환율 값
-}
