+++ ExchangesApp/ViewController.swift
@@ -1,0 +1,19 @@
+//
+//  ViewController.swift
+//  ExchangesApp
+//
+//  Created by 김우성 on 7/7/25.
+//
+
+import UIKit
+
+class ViewController: UIViewController {
+
+  override func viewDidLoad() {
+    super.viewDidLoad()
+    print("viewDidLoad")
+  }
+
+
+}
+
